==> scripts/01-download_data.R <==
#### Preamble ####
# Purpose: Downloads and saves the data from Baseball Reference.
# Author: Tracy Yang
# Date: 6 April 2024
# Contact: ycart.yang@mail.utoronto.ca
# License: MIT
# Pre-requisites: -


#### Download data ####
# Team data downloaded from https://www.baseball-reference.com/leagues/ and then 
# converted from xls files to csv files to be saved in the raw data folder.
# Hitting, pitching, and fielding data downloaded separately for each year to be combined later.
# 2000 - 2010 data were downloaded for model and 2024 data for prediction, in total 36 files.
